SmtpClient minor refactoring
